third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/gen-reconnect-during.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let device, fake_peripheral, service, fake_service;
  return getHealthThermometerDeviceWithServicesDiscovered({
      filters: [{services: [health_thermometer.name]}]
  })
    .then(_ => ({device, fake_peripheral} = _))
    .then(() => fake_peripheral.addFakeService({uuid: health_thermometer.name}))
    .then(s => (fake_service = s))
    .then(() => device.gatt.getPrimaryService('health_thermometer'))
    .then(s => service = s)
    .then(() => {
      let promise = assert_promise_rejects_with_message(service.getCharacteristic('heart_rate_measurement'), new DOMException(
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'));
      device.gatt.disconnect();
      return device.gatt.connect().then(() => promise);
    });
}, 'disconnect() and connect() called during getCharacteristic. Reject with ' +
   'NetworkError.');

</script>
